Add 'Request new item' template button in RequestItemView
The RequestItemView now has a static method that returns an Item
object with the 'img_url' pointing to a request image.  The image is
currently pointing to a random image I found online and I still need
to make an image that says 'Can't find something?  Request an image here'