Add feature to request new items
The meteor server now adds requested items via the iOS app to a Requests
collection.  I patched the AFNetworking library to work well with POST
requests (the library wasn't working right in terms of serializing the
data as a good string).  The entry point for the Item models are implemented
as a static function in the Item model.